add commit

+++ database/migrations/2022_10_05_082954_create_s_p_p_d_s_table.php
@@ -18,7 +18,7 @@ class CreateSPPDSTable extends Migration
             $table->string('no_sppd')->unique();
             $table->string('doksebelumnya')->nullable();
             $table->date('tglmasuk')->nullable();
-            $table->float('jumlah')->nullable();
+            $table->integer('jumlah')->nullable();
             $table->timestamps();
         });
     }
